+++ src/main/java/org/sopt/diary/api/DiaryController.java
@@ -2,7 +2,13 @@ package org.sopt.diary.api;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
-import java.util.Objects;
+import org.sopt.diary.api.request.DiaryRequest;
+import org.sopt.diary.api.response.CommonDiaryListResponse;
+import org.sopt.diary.api.response.CommonDiaryResponse;
+import org.sopt.diary.api.response.DetailDiaryResponse;
+import org.sopt.diary.api.response.DiaryListResponse;
+import org.sopt.diary.api.response.MyDiaryListResponse;
+import org.sopt.diary.api.response.MyDiaryResponse;
 import org.sopt.diary.domain.Diary;
 import org.sopt.diary.exception.NotFoundUserException;
 import org.sopt.diary.repository.constant.Category;
