+++ src/main/java/org/sopt/diary/api/request/DiaryCategoryRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.diary.api;
+package org.sopt.diary.api.request;
 
 public record DiaryCategoryRequest(String title, String content, String category) {
