+++ src/main/java/org/sopt/diary/domain/User.java
@@ -1,7 +1,6 @@
 package org.sopt.diary.domain;
 
 
-import java.time.LocalDateTime;
 import java.util.List;
 import org.sopt.diary.repository.UserEntity;
 
@@ -11,25 +10,22 @@ public class User {
 
     private String password;
 
-    private List<Diary> diaryEntities;
-
     private String nickname;
 
-    public User(String userName, String password, List<Diary> diaryEntities,
+    public User(String userName, String password,
         String nickname) {
         this.userName = userName;
         this.password = password;
-        this.diaryEntities = diaryEntities;
         this.nickname = nickname;
     }
 
     public static User from(UserEntity userEntity) {
-        return new User(userEntity.getUserName(), userEntity.getPassword(), userEntity.getDiaryEntities().stream().map(Diary::of).toList(),
+        return new User(userEntity.getUsername(), userEntity.getPassword(),
             userEntity.getNickname());
     }
 
     public static UserEntity toEntity(User user){
-        return new UserEntity(user.getUserName(), user.getPassword(), user.getDiaryEntities().stream().map(Diary::toEntity).toList(),
+        return new UserEntity(user.getUserName(), user.getPassword(),
             user.getNickname());
 
     }
@@ -42,9 +38,6 @@ public class User {
         return password;
     }
 
-    public List<Diary> getDiaryEntities() {
-        return diaryEntities;
-    }
 
     public String getNickname() {
         return nickname;
